last steps for the release
 - remove the explizit 'true' value from the travis used spec tests
 - rewrote users Documentation
 - seperate the Client Class into an own File
 - use 'pre' as Version postfix
