+++ .travis.yml
@@ -32,7 +32,7 @@ before_install:
   - docker ps
   - sleep 20s
 
-script: bundle exec rspec spec || true
+script: bundle exec rspec spec
 
 fast_finish: true
