+++ lib/icinga2/version.rb
@@ -3,20 +3,7 @@
 
 module Icinga2
 
-  # namespace for version information
-  module Version
-
-    # major part of version
-    MAJOR = 0
-    # minor part of version
-    MINOR = 9
-    # tiny part of version
-    TINY  = 95
-    # patch part
-    PATCH = 5
-  end
-
   # Current version of gem.
-  VERSION = [Version::MAJOR, Version::MINOR, Version::TINY, Version::PATCH].compact * '.'
+  VERSION = '1.1.0.pre1' # [Version::MAJOR, Version::MINOR, Version::TINY, Version::PATCH].compact * '.'
 
 end
